Module Yuujinchou.Pattern

type path = string list

The type of names.

type 'a pattern

The type of patterns, parametrized by the attribute type.

val inv : 'a pattern -> 'a pattern

Inverting the pattern.

val wildcard : 'a pattern

Wildcard pattern.

val root : 'a pattern

The pattern that only matches the root.

val scope : path -> 'a pattern -> 'a pattern

Scoping a pattern.

val renaming_scope : path -> path -> 'a pattern -> 'a pattern

Scoping a pattern and rename the prefix.

val seq : 'a pattern list -> 'a pattern

Sequencing.

val none : 'a pattern

The pattern that matches nothing.

val any : 'a pattern

The pattern that matches any name.

val id : path -> 'a pattern

The pattern that matches only the specified name.

val renaming : path -> path -> 'a pattern

The pattern that matches only the specified name and replaces it.

val prefix : path -> 'a pattern

The pattern that matches any name with the given prefix.

val renaming_prefix : path -> path -> 'a pattern

The pattern that matches any name with the given prefix and replaces the prefix.

val attr : 'a -> 'a pattern -> 'a pattern

The pattern that assigns the default attribute.

val join : 'a pattern list -> 'a pattern

Join of a list of patterns.

val meet : 'a pattern list -> 'a pattern

Meet of a list of patterns.

val skip : path -> 'a pattern

The pattern that skips the specified name.

val skip_prefix : path -> 'a pattern

The pattern that skips any name with the specified prefix.

val pp_path : Stdlib.Format.formatter -> path -> unit

Pretty-printer for path.

val pp_pattern : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a pattern -> unit

Pretty-printer for pattern.