Parameter Make.Param

type data

The type of data held by the bindings. The difference between data and tags is that the data will survive the efficient retagging. See Trie.retag.

type tag

The type of tags attached to the bindings. The difference between data and tags is that tags can be efficiently reset. See Trie.retag.

type hook

The type of modifier hook labels. This is for extending the modifier language.

type context

The type of contexts passed to each call of Modifier.S.modify for the effect handler to distinguish different function calls.