Module Yuujinchou__Trie

Types

type path = string list

The type of hierarchical names. The name x.y.z is represented by the OCaml list ["x"; "y"; "z"].

type +'a t

The abstract type of a trie.

Basic Operations

val empty : 'a t

The empty trie.

val is_empty : 'a t -> bool

Check whether the trie is empty.

val mk_root : 'a option -> 'a t

Make a trie with only one binding: the root associated with the provided value. mk_root None will make an empty trie and mk_root (Some v) will make a trie with the value v. If thi input is always Some v, use root.

val prefix : path -> 'a t -> 'a t

prefix p t makes a minimum trie with t rooted at p.

val singleton : (path * 'a) -> 'a t

strington (p, d) makes a trie with the only binding: p and its associated value d.

val root : 'a -> 'a t

root d makes a trie with the only binding: the root and its associated value d. It is equivalent to singleton [] d.

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

equal eq t1 t2 checks whether two tries are equal. It the internal representations of tries are physically equal, equal eq t1 t2 will return true without calling eq.

Locating Values

val find_subtree : path -> 'a t -> 'a t

find_subtree p t returns the subtree rooted at p.

val find_singleton : path -> 'a t -> 'a option

find_singleton p t returns the value at p.

val find_root : 'a t -> 'a option

find_root t returns the value at the root. This is equivalent to find_singleton [] t.

Mapping and Filtering

val filter_map_endo : ('a -> 'a option) -> 'a t -> 'a t

filter_map_endo f t applies the function f to each value v in the trie. If f v returns None, then the binding will be removed from the trie. Otherwise, if f v returns Some v', then the value will be replaced by v' in the returned trie.

Union

val union : ('a -> 'a -> 'a) -> 'a t -> 'a t -> 'a t

union merger t1 t2 merges two tries t1 and t2. If both tries have a binding at the same path, it will call the function merger to reconcile the values from the two tries.

val union_subtree : ('a -> 'a -> 'a) -> 'a t -> (path * 'a t) -> 'a t

union_subtree merger t1 (path, t2) is equivalent to union merger t1 (prefix t2), but potentially more efficient.

Separation

val detach_subtree : path -> 'a t -> 'a t * 'a t

detach p t detaches the subtree at p from the main trie and returns both the subtree and the remaining trie. If detach p t returns t1, t2, then union_subtree m t2 (p, t1) should be equivalent to t.

val detach_singleton : path -> 'a t -> 'a option * 'a t

detach p t detaches the binding at p from the main trie and returns both the binding and the remaining trie. If detach p t returns b, t', then union_subtree m t' (p, mk_root b) should be equivalent to t.

Iterators

val to_seq : 'a t -> (path * 'a) Stdlib.Seq.t

to_seq t traverses through the trie t in the lexicographical order.

val to_seq_values : 'a t -> 'a Stdlib.Seq.t

to_seq_values t traverses through the trie t in the lexicographical order but only returns the associated values. This is faster than Seq.map snd @@ to_seq t because it does not need to reconstruct the paths.

val of_seq : ('a -> 'a -> 'a) -> (path * 'a) Stdlib.Seq.t -> 'a t

of_seq m s inserts bindings (p, d) into an empty trie, one by one, using union_subtree.

Pretty Printer

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

pp pp_v t prints out the content of t, using the pretty printer pp_v on values.

Physical Equality

val physically_equal : 'a t -> 'a t -> bool

This is an internal API for testing whether the library is preserving physical equality as much as possible. If physically_equal t1 t2 returns true then equal eq t1 t2 must return true. Do not rely on this function unless you know the internals of tries.