Module Yuujinchou

Yuujinchou is an OCaml package of name patterns. It was motivated by the "import" or "include" statements present in almost all programming languages. Here are a few examples:

open import M -- Agda

import foo # Python

The ability to import content from other files helps organize code. However, it also poses a new challenge: how could programmers prevent imported content from shadowing existing content? For example, if we already have a function test in the current scope, maybe we do not wish to import another function also named test. To address this, many programming languages allow programmers to selectively hide or rename part of the imported content:

open import M renaming (a to b) public
-- renaming a to b, and then re-exporting the content

import foo as bar
# putting content of foo under the prefix bar

We can view a collection of hierarchical names as a tree, and see these modifiers as tree transformers. I took this aspect seriously and designed a powerful (possibly overkilling) combinator calculus to express such tree transformers---the library you are checking now. It supports renaming, scopes, sequencing, unions, generic filtering.

Organization

The code is split into three parts:

module Trie : module type of Trie

The Trie module implements a data structure (tries) that supports efficient subtree operators on a collection of hierarchical names and their associated data.

module Pattern : sig ... end

The Pattern module defines the patterns.

module Action : sig ... end

The Action module implements the engine running the patterns.

How to Use It

open Yuujinchou

module Data =
struct
  type t = int
  let equal n1 n2 = n1 = n2
  let merge x y =
    if equal x y then x
    else failwith "Inconsistent data assigned to the same path."
  let shadow _x y = y
  let compare : t -> t -> int = compare
end

(** An environment is a mapping from paths to data. *)
type env = Data.t Trie.t

(** [remap pattern env] uses the [pattern] to massage
    the environment [env]. *)
let remap pattern env =
  let pp_path = function [] -> "(root)" | path -> String.concat "." path in
  match Action.run Data.merge pattern env with
  | Ok env -> env
  | Error (Action.BindingNotFound path) ->
    failwith ("Expected binding(s) not found within the subtree at " ^ pp_path path ^ ".")

(** [import env pattern imported] imports the environment
    [imported] massaged by [pattern] into [env]. *)
let import env pattern imported =
  Trie.union Data.shadow env @@ remap pattern imported

module DataSet = Set.Make (Data)

(** [select env pattern] returns the set of matched data. *)
let select env pattern =
  DataSet.of_seq @@ Trie.to_seq_values @@ remap pattern env

Namespace?

This library intends to treat a namespace as the prefix of a group of names. That is, there is no namespace a, but only a group of unrelated names that happen to have the prefix a.

Note that namespaces (name prefixes of unrelated items) are different from modules (groups of items that are bound together). This library does not provide special support for modules (yet).

Examples from Other Languages

Haskell

import Mod -- x is available an both x and Mod.x

union [any; renaming_subtree [] ["Mod"]]

import Mod (x,y)

union [only ["x"]; only ["y"]]

import qualified Mod

renaming_subtree [] ["Mod"]

import qualified Mod hiding (x,y)

seq [except ["x"]; except ["y"]; renaming_subtree [] ["Mod"]]

Racket

(require (only-in ... id0 [old-id1 new-id1]))

seq [...; union [only ["id0"]; seq [only ["old-id1"]; renaming ["old-id1"] ["new-id1"]]]]

(require (except-in ... id0 id1]))

seq [...; except ["id0"]; except ["id1"]]

(require (prefix-in p: ...))

seq [...; renaming_subtree [] ["p"]]

(require (rename-in ... [old-id0 new-id0] [old-id1 new-id1]))

seq [...; renaming ["old-id0"] ["new-id0"]; renaming ["old-id1"] ["new-id1"]]

(require (combine-in require-spec0 require-spec1 ...))

union [require-spec0; require-spec1; ...]

The provide mechanism can be implemented in a similar way, and the phases can be handled via Pattern.filter_map.

What is "Yuujinchou"?

"Yuujinchou" is the transliteration of "友人帳" in Japanese, which literally means "book of friends". It is a powerful notebook in the manga Natsume Yuujinchou (夏目友人帳) that collects many real names (真名) of youkais (妖怪) (supernatural and spiritual monsters). These real names can be used to summon and control youkais, but the protagonist decided to return the names to their original owners. The plot is about meeting all kinds of youkais.

This magical book will automatically turn to the page with the correct name when the protagonist pictures the youkai in his mind. This library is also about finding real names of youkais.

The transliteration is in the Wāpuro style to use only English alphabet letters; otherwise, its Hepburn romanization would be "Yūjin-chō".