Module Yuujinchou

module Pattern : sig ... end

The Pattern module defines the patterns.

module Action : sig ... end

The Action module implements the engine runnig the patterns.

Introduction

Yuujinchou is an OCaml combinator library for manipulating names. It was motivated by the `import` or `include` statements present in almost all programming languages. Here are some examples of such statements:

open import M -- Agda

import foo # Python

The ability to import content from other files helps organize code. However, it also brings up new design issues: how could programmers prevent imported content from colliding or shadowing the existing content in the current scope? For example, maybe in the current scope, we have defined a function called test, and do not want to import another function also called test. To address this, many programming languages allow programmers to selectively hide or rename part of the imported content:

open import M renaming (a to b) public
-- renaming a to b, and then re-exporting the content

import foo as bar
# putting content of foo under the prefix bar

We can treat hiding or renaming as a partial function from names to names. I took this aspect seriously and designed this combinator library---a concise yet powerful (perhaps overkilling) language for manipulating names in programming languages.

Applications

This library was motivated by the import mechanism in most programming languages, but can be used in any situation involving selecting names. For example, during interactive theorem proving, perhaps you want to unfold some definitions but not others. This library gives you a powerful language to support fancy name selection.

Organization

The code is split into two parts: Pattern and Action.

Namespace Support

This library intends to treat a namespace as the prefix of a group of names. That is, there is no namespace a, but only a group of unrelated names that happen to have the prefix a.

Note that namespaces (name prefixes of unrelated declarations) are different from modules (groups of declarations that are bound together). This library does not provide special support for modules (yet).

Examples

Haskell

import Mod -- x is available an both x and Mod.x

join [any; renaming_prefix [] ["Mod"]]

import Mod (x,y)

join [id ["x"]; id ["y"]]

import qualified Mod

join [renaming_prefix [] ["Mod"]]

import qualified Mod hiding (x,y)

renaming_scope [] ["Mod"] @@ meet [hide ["x"]; hide ["y"]]

What is "Yuujinchou"?

"Yuujinchou" is the transliteration of "友人帳" in Japanese, which literally means "book of friends". The book is a powerful booklet that collects real names (真名) of youkais (妖怪) (supernatural and spiritual monsters) in the manga and anime Natsume Yuujinchou (夏目友人帳). These real names can be used to control youkais, but the protagonist decided to return the names to youkais. The plot is about meeting different youkais during the journey.

This library is also about using names to summon monsters.

The transliteration is in the Wāpuro style to work around the name restriction; otherwise, its Hepburn romanization would be "Yūjin-chō".