Module Yuujinchou

yuujinchou is an OCaml package of name modifiers.

module Trie : sig ... end

The Trie module implements mappings from paths to values that support efficient subtree operations.

module Language : sig ... end

The Language module defines the language of modifiers.

module Modifier : sig ... end

The Modifier module implements the engine running the modifiers of type Language.t.

module Scope : sig ... end

The Scope module implements lexical scoping based on Modifier.