Module Yuujinchou.Action

The Action module implements the engine running the patterns.

Matching

val run : ?⁠rev_prefix:Pattern.path -> union:(rev_path:Pattern.path -> 'a -> 'a -> 'a) -> unit Pattern.t -> 'a Trie.t -> ('a Trie.t, [> `BindingNotFound of Pattern.path ]) Stdlib.result

run ~rev_prefix ~union pattern trie runs the pattern on the trie and return the transformed trie.

parameter rev_prefix

The prefix prepended to any path sent to union and any path in the error reporting, but in reverse. The default is the empty unit path ([]).

parameter union

The resolver for two conflicting bindings sharing the same name. Patterns such as Pattern.renaming and Pattern.union could lead to conflicting bindings, and union ~rev_path x y should return the resolution of x and y at the (reversed) path rev_path.

returns

The new trie after the transformation. Error (`BindingNotFound p) means the transformation failed because of the absence of expected bindings. For example, the pattern Pattern.except_subtree["x"; "y"] expects that there was already something under the subtree at x.y. If there were actually no names with the prefix x.y, then the pattern will trigger the error `BindingNotFound ["x"; "y"]. The path is only an approximation---the user might have intended to hide the binding at ["x"; "y"; "z"], a binding under ["x"; "y"], but the engine would never know the user's true intension.

val run_with_custom : ?⁠rev_prefix:Pattern.path -> union:(rev_path:Pattern.path -> 'a -> 'a -> 'a) -> custom:(rev_path:Pattern.path -> 'custom -> 'a -> 'a option) -> 'custom Pattern.t -> 'a Trie.t -> ('a Trie.t, [> `BindingNotFound of Pattern.path ]) Stdlib.result

run_with_custom ~rev_prefix ~custom ~union pattern trie runs the pattern on the trie and return the transformed trie. It is similar to run but accepts a new argument custom.

parameter custom

The customer filter that will be triggered by the pattern Pattern.customf. When the engine encounters Pattern.customf, it will call custom ~rev_path:p f v for all data v at p in the trie.

Pretty Printers

val pp_path : Stdlib.Format.formatter -> Pattern.path -> unit

Pretty printer for path.