Module Yuujinchou.Action

The Action module implements the engine running the patterns.

Error Type

type error =

| Binding_not_found of Pattern.path
The engine could not find the expected bindings.

The type of errors due to the absense of expected bindings. For example, the pattern Pattern.except_subtree ["x"; "y"] expects that there was already something under the subtree at x.y. If there were no names with the prefix x.y, then the pattern will result into the error Binding_not_found ["x"; "y"].

Matching

val run : ('a -> 'a -> 'a) -> 'a Pattern.t -> 'a Trie.t -> ('a Trie.t, error) Stdlib.result

run merger pattern trie runs the pattern on the trie

parameter merger

The resolver for two conflicting bindings sharing the same name. Patterns such as Pattern.renaming and Pattern.union could lead to conflicting bindings, and merger x y should return the resolution of x and y.

Pretty Printers

val pp_error : Stdlib.Format.formatter -> error -> unit

Pretty printer for error.