Module Yuujinchou.Action

The Action module implements the engine runnig the patterns.

Types

type 'a result_ = [

| `NoMatch
The pattern does not match the name.

| `Matched of (Pattern.path * 'a) list
The pattern matches the name, with a list of tagged new names.

]

The result type of pattern matching.

type 'a error =

| ReplacementNotUsed of 'a Pattern.pattern
Renaming patterns are run under the inverse mode.

| EmptyMeet of 'a Pattern.pattern
The join patterns under the inverse mode (or, equivalently, the meet patterns under the normal mode) have no subpatterns.

The type of errors due to the violation of some invariant. See Invariants. It should be impossible to violate these invariants unless Pattern.unsafe_meet or Pattern.unsafe_inv is used.

The pattern embedded in the error message is the fragment that violates the invariant. The pattern pat in EmptyMeet pat is not useful on its own---we all know it must be PatJoin []---but it facilitates using or-patterns in error handling.

Engines

val run : default:'a -> join:('a -> 'a -> 'a) -> meet:('a -> 'a -> 'a) -> 'a Pattern.pattern -> Pattern.path -> ('a result_, 'a error) Stdlib.result

The pattern matching engine.

parameter default

The default attribute for the engine to start with. See Attributes.

parameter join

The join operator to resolve conflicting attributes. See Attributes.

parameter meet

The meet operator to resolve conflicting attributes. See Attributes.

val run_ : unit Pattern.pattern -> Pattern.path -> (unit result_, unit error) Stdlib.result

The engine specialized to unit pattern where the attribute type is unit.

val check : 'a Pattern.pattern -> (unit, 'a error) Stdlib.result

Check whether a pattern violates any invariant. See Invariants.

Pretty Printers

val pp_result_ : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a result_ -> unit

Pretty printer for result_.

val pp_result : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> ('a result_, 'a error) Stdlib.result -> unit

Pretty printer for results of pattern matching.

val pp_check_result : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> (unit, 'a error) Stdlib.result -> unit

Pretty printer for results of checking patterns.