Module Yuujinchou.Pattern

The Pattern module defines the patterns.

Pattern Type

type 'a t

The type of patterns, parametrized by the type of associated data.

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

Checking equality. Note that patterns created by filter_map are not comparable to each other because they take a function.

Hierarchical Names

type path = string list

The type of names.

Pattern Builders

Basics

val any : 'a t

any keeps the content of the current tree. It is an error if the tree is empty (no name to match).

val root : 'a t

root keeps only the empty name (the empty list []). It is equivalent to only [].

val wildcard : 'a t

wildcard keeps everything except the empty name (the empty list []). It is an error if there was no name is matched.

val only : path -> 'a t

only x keeps the name x and drops everything else. It is an error if there was no binding named x in the tree.

val only_subtree : path -> 'a t

only_subtree path keeps the subtree rooted at path. It is an error if the subtree was empty.

val in_subtree : path -> 'a t -> 'a t

in_subtree path pattern runs the pattern pat on the subtree rooted at path. Bindings outside the subtree are kept intact.

Negation

val none : 'a t

none drops everything. It is an error if the tree was already empty (nothing to drop).

val except : path -> 'a t

except x drops the binding at x. It is an error if there was no x from the beginning.

val except_subtree : path -> 'a t

except_subtree p drops the subtree rooted at p. It is an error if there was nothing in the subtree. This is equivalent to in_subtree p none.

Renaming

val renaming : path -> path -> 'a t

renaming x x' renames the binding at x to x'. Note that such renaming does not affect names under x. See renaming_subtree for comparison. It is an error if there was no x from the beginning.

val renaming_subtree : path -> path -> 'a t

renaming_subtree path path' relocates the subtree rooted at path to path'. If you only want to move the root, not the entire subtree, see renaming. It is an error if the subtree was empty (nothing to move).

Associated Data

val filter_map : ('a -> 'a option) -> 'a t

filter_map f applies the function f to the associated data in the tree. If the result is None, the binding would be removed from the tree. Otherwise, if the result is Some x, the content of the binding would be replaced with x.

Sequencing

val seq : 'a t list -> 'a t

seq [pat0; pat1; pat2; ...; patn] runs the patterns pat0, pat1, pat2, ..., patn in order.

Union

val union : 'a t list -> 'a t

union [pat0; pat1; pat2; ...; patn] calculates the union of the results of individual patterns pat0, pat1, pat2, ..., patn.

Pretty Printers

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

Pretty printer for t.